Docker dev2

+++ server/dataset/termdb.test.js
@@ -158,6 +158,8 @@ if (!fs.existsSync(datadir)) fs.mkdirSync(datadir, { recursive: true }) // creat
 
 const srcdb = path.join(serverconfig.binpath, 'test/testdata/db2')
 const destdb = path.join(serverconfig.tpmasterdir, ds.cohort.db.file)
-fs.copyFileSync(srcdb, destdb)
+if (!fs.existsSync(destdb) && !process.env.PP_MODE?.startsWith('container-')) {
+	fs.copyFileSync(srcdb, destdb)
+}
 
 module.exports = ds
